# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Navigation
  description: "Navigation"
host: localhost:10010
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
tags:
  - name: 'Register'
    description: Felhasználói regisztrációs műveletek
  - name: 'Authentication'
    description: Felhasználói autentikációs műveletek
  - name: 'Distance'
    description: Távolság műveletek
  - name: 'Location'
    description: Hely műveletek
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: x_session_id
security:
  - ApiKeyAuth: []
paths:
  /signUp:
    x-swagger-router-controller: signUp
    post:
      description: Új felhasználó készités
      operationId: createUser
      security: []
      tags:
      - "Register"
      parameters:
        - name: user
          in: body
          description: Létrehozás
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Sikeres
          schema:
            $ref: "#/definitions/CreateUserResponse"
        default:
          description: Hiba
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authentication/login:
    x-swagger-router-controller: authentication
    post:
      description: Felhasználó hitelesités
      operationId: login
      security: []
      tags:
      - "Authentication"
      parameters:
        - name: credentials
          in: body
          description: Hitelesítő adatok
          required: true
          schema:
            $ref: "#/definitions/Credentials"
      responses:
        "200":
          description: Sikeres
          schema:
            $ref: "#/definitions/AuthenticationResponse"
        "401":
          description: Nem jogos
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Hiba
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authentication/logout:
    x-swagger-router-controller: authentication
    get:
      description: Kijelentkezés
      operationId: logout
      tags:
      - "Authentication"
      parameters:
        - name: x_session_id
          in: header
          description: Session ID
          required: true
          type: string
      responses:
        "200":
          description: Sikeres
          schema:
            $ref: "#/definitions/AuthenticationResponse"
        "401":
          description: Nem jogos
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Hiba
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
  /distance:
    x-swagger-router-controller: distance
    get:
      description: Két helység távolság.
      operationId: distance
      tags:
        - "Distance"
      parameters:
        - name: origin
          in: query
          description: Forrás
          required: false
          type: string
        - name: destination
          in: query
          description: Cél
          required: true
          type: string
      responses:
        "200":
          description: Sikeres
          schema:
            $ref: "#/definitions/DistanceResponse"
        "400":
          description: Rossz kérés
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Hiba
          schema:
            $ref: "#/definitions/ErrorResponse"
  /location/setLocation:
    x-swagger-router-controller: location
    put:
      description: Helység frissités
      operationId: setLocation
      tags:
        - "Location"
      parameters:
        - name: x_session_id
          in: header
          description: Session ID
          required: true
          type: string
        - name: location
          in: query
          description: Jelenlegi hely
          required: true
          type: string
      responses:
        "200":
          description: Sikeres
          schema:
            $ref: "#/definitions/LocationResponse"
        "400":
          description: Nem sikerült
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Hiba
          schema:
            $ref: "#/definitions/ErrorResponse"
  /location/deleteLocation:
    x-swagger-router-controller: location
    delete:
      description: Jelenlegi hely törlés
      operationId: deleteLocation
      tags:
        - "Location"
      parameters:
        - name: x_session_id
          in: header
          description: Session ID
          required: true
          type: string
      responses:
        "200":
          description: Sikeres
          schema:
            $ref: "#/definitions/LocationResponse"
        "400":
          description: Törlés sikertelen
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Hiba
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  User:
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: 'yourname'
      password:
        type: string
        example: 'password'
  CreateUserResponse:
    properties:
      _id:
        type: string
        example: 'somestring'
  Credentials:
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: 'yourname'
      password:
        type: string
        example: 'password'
  AuthenticationResponse:
    required:
      - sessionID
    properties:
      sessionID:
        type: string
        example: 'somestring'
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  DistanceResponse:
    required:
      - distance
    properties:
      distance:
        type: string
        example: '130 km'
  LocationResponse:
    required:
      - location
    properties:
      location:
        type: string
        example: 'Szeged'